feat: add more coverage
