+++ tests/fixtures/deployments/0.contracts.json
@@ -1,0 +1,1 @@
+[]
