+++ tests/library/mod.rs
@@ -1,1 +1,1 @@
-mod accounts;
+mod deployments;
